backend/requirements.txt
asgiref==3.9.1
certifi==2025.8.3
cffi>=1.17.1
charset-normalizer==3.4.3
cryptography==45.0.6
defusedxml==0.7.1
dj-database-url==3.0.1
Django==5.2.5
django-cors-headers==4.7.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
djoser==2.2.3
gunicorn==23.0.0
idna==3.10
oauthlib==3.3.1
packaging==25.0
pillow==11.3.0
psycopg2-binary>=2.9.10
pycparser==2.22
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.1.1
python3-openid==3.2.0
pytz==2025.2
requests==2.32.5
requests-oauthlib==2.0.0
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.9.0
beautifulsoup4==4.12.3
extruct==0.16.0
lxml>=4.9.3
w3lib==2.2.1
selectolax>=0.3.25
tldextract==5.1.3
pycountry==24.6.1
rapidfuzz>=3.10.1
requests-cache==1.2.1
tenacity==9.0.0
pydantic>=2.10.4
price-parser==0.3.4
scrapegraph_py==1.20.0
celery==5.3.6
django-celery-beat==2.8.0
django-celery-results==2.5.1
redis==5.2.1
